requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.2.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
djangorestframework-simplejwt>=5.2.2
Pillow>=10.0.0
